+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Studio : group by application on json model grid view.
 - Export studio app: email action - email template
 - Export Studio app: export actions created with meta-action-from
+- JSON-MODEL-FORM : add tracking on json fields
 
 ## Improvements
 - Remove Pack Feature from Product.